todolistproject/requirements.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==44.0.2
Django==5.1.7
django-environ==0.12.0
djangorestframework==3.15.2
pi==0.1.2
psycopg2-binary==2.9.10
pycparser==2.22
pycryptodome==3.22.0
PyJWT==2.10.1
python-decouple==3.8
python-dotenv==1.0.1
sqlparse==0.5.3
tzdata==2025.1
